Move useWithdraw out of its folder
