--- src/pages/Admin/Charity/Withdraws/Withdrawer/WithdrawForm/useWithdraw/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./useWithdraw";
